requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
psycopg2==2.9.9
Werkzeug==3.0.1
python-dotenv==1.0.0 